+++ features/sessions/ui/src/commonMain/kotlin/app/campfire/sessions/ui/playback/expanded/ExpandedPlaybackBar.kt
@@ -576,7 +576,7 @@ private fun SharedTransitionScope.ExpandedPlaybackContent(
           playbackSpeed = playerState.speed,
           onClick = {
             scope.launch {
-              overlayHost.showPlaybackSpeedBottomSheet(playerState.speed)
+              overlayHost.showPlaybackSpeedBottomSheet(session!!.libraryItem.id, playerState.speed)
             }
           },
         )