+++ features/sessions/ui/src/commonMain/kotlin/app/campfire/sessions/ui/playback/expanded/SmallExpandedPlaybackBar.kt
@@ -618,7 +618,7 @@ private fun PlaybackOptionsColumn(
         playbackSpeed = playerState.speed,
         onClick = {
           scope.launch {
-            overlayHost.showPlaybackSpeedBottomSheet(playerState.speed)
+            overlayHost.showPlaybackSpeedBottomSheet(session!!.libraryItem.id, playerState.speed)
           }
         },
       )